feat split models avoid python imports cycle issues
