+++ nectar_comm/models/base/client.py
@@ -1,0 +1,5 @@
+from models.base.nectar_base_model import NectarBaseModel, Named
+
+
+class ClientBaseModel(Named, NectarBaseModel):
+    pass
