+++ nectar_comm/models/base/clinic.py
@@ -1,0 +1,9 @@
+from models.base.nectar_base_model import NectarBaseModel, Named
+
+class ClinicBaseModel(Named, NectarBaseModel):
+    pass
+
+
+
+
+
