+++ nectar_comm/models/base/employee.py
@@ -1,0 +1,6 @@
+from models.base.nectar_base_model import NectarBaseModel, Named
+
+
+class EmployeeBaseModel(Named, NectarBaseModel):
+    pass
+
