+++ nectar_comm/models/base/patient.py
@@ -1,0 +1,6 @@
+from html.entities import name2codepoint
+from models.base.nectar_base_model import NectarBaseModel, Named
+
+
+class PatientBaseModel(Named, NectarBaseModel):
+    pass
