+++ nectar_comm/models/nectar_relationship_models.py
@@ -1,11 +1,14 @@
 from tortoise import fields
-from models.base.nectar_base_model import (
-    ClientBaseModel,
-    CommunicationBaseModel,
-    PatientBaseModel,
-    EmployeeBaseModel,
-    ClinicBaseModel
-)
+from models.base.client import ClientBaseModel
+from models.base.clinic import ClinicBaseModel
+from models.base.employee import EmployeeBaseModel
+from models.base.patient import PatientBaseModel
+from models.base.communication import CommunicationBaseModel
+
+
+"""
+    Note: Working the relations in the same file allows to avoid the problems with the import cycle in python
+"""
 
 class Clinic(ClinicBaseModel):
     communications: fields.ReverseRelation["Communication"]
@@ -36,5 +39,3 @@ class Communication(CommunicationBaseModel):
     patient: fields.ForeignKeyRelation[Patient] = fields.ForeignKeyField(
         "models.Patient", related_name="communications", to_field="id"
     )
-
-
