+++ nectar_comm/models/query/client.py
@@ -1,0 +1,7 @@
+from models.nectar_relationship_models import Client as Relations
+
+
+class Client(Relations):
+    class Meta:
+        table = "clients"
+        table_description = "Saved all info about the client"
