+++ nectar_comm/models/query/clinic.py
@@ -1,0 +1,7 @@
+from models.nectar_relationship_models import Clinic as Relations
+
+
+class Clinic(Relations):
+    class Meta:
+        table = "clinics"
+        table_description = "Saved all info about the clinic"
