+++ nectar_comm/models/query/communication.py
@@ -1,0 +1,7 @@
+from models.nectar_relationship_models import Communication as Relations
+
+
+class Communication(Relations):
+    class Meta:
+        table = "communications"
+        table_description = "Saved all info and type of contact with the patient"
