+++ nectar_comm/models/query/employee.py
@@ -1,0 +1,7 @@
+from models.nectar_relationship_models import Employee as Relations
+
+
+class Employee(Relations):
+    class Meta:
+        table = "employees"
+        table_description = "Saved all info about the employee"
